Add versioning bits from DWS.
Trigger workflows only on push.  The 'pull_request' trigger was redundant and
caused a PR to get two builds.

Add git-version-gen tool. Modify Makefile to allow .version to be generated.

Remove the DWS controller node selector, cray.wlm.manager.  This makes it
easier to test with modern DWS which doesn't use that.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>