Dynamically create an overlay for the image tag
Avoid editing config/manager/kustomization.yaml to set the image tag.  This
causes git to consider the workarea to be dirty, so two consecutive deploys
from a workarea that has no other changes will result in the second deploy
looking for an image with a "-dirty" tag.

Instead, from the makefile we create a throw-away, and untracked, overlay that
will be used to set the image tag

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>